Renamed shared-components to components, put each shared component in a subdirectory.
