+++ assets/components/infoText/infoText.scss
