+++ assets/stylesheets/main.scss
@@ -7,10 +7,10 @@
 
 // ----- Shared Components ----- //
 
-@import '../shared-components/doubleHeading';
-@import '../shared-components/featureList';
-@import '../shared-components/infoText';
-@import '../shared-components/ctaLink';
+@import '../components/doubleHeading/doubleHeading';
+@import '../components/featureList/featureList';
+@import '../components/infoText/infoText';
+@import '../components/ctaLink/ctaLink';
 
 
 // ----- Pages ----- //
